kg-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.3
httpx==0.25.2
python-multipart==0.0.6
neo4j==5.14.1
sentence-transformers==2.2.2
spacy==3.7.2
numpy==1.24.3
torch==2.1.0
PyYAML==6.0.1
python-dateutil==2.8.2
tiktoken==0.5.1
